* UI made great again

+++ apps/provis/Camera_View.cpp
@@ -13,16 +13,16 @@ void Camera_View::translate(scm::math::vec3f offset)
     _position += offset;
     update_matrices();
 }
-void Camera_View::update_window_size(int width_window, int height_window)
+void Camera_View::update_window_size(float width_window, float height_window)
 {
     _width_window = width_window;
     _height_window = height_window;
 
-    _matrix_perspective = scm::math::make_perspective_matrix(_fov, float(_width_window) / float(_height_window), _plane_near, _plane_far);
+    _matrix_perspective = scm::math::make_perspective_matrix(_fov, _width_window / _height_window, _plane_near, _plane_far);
 
     lamure::ren::policy *policy = lamure::ren::policy::get_instance();
-    policy->set_window_width(_width_window);
-    policy->set_window_height(_height_window);
+    policy->set_window_width((const int32_t) _width_window);
+    policy->set_window_height((const int32_t) _height_window);
 }
 void Camera_View::update_matrices()
 {
@@ -96,8 +96,8 @@ scm::math::mat4f &Camera_View::get_matrix_view()
     }
     return _matrix_view;
 }
-int &Camera_View::get_width_window() { return _width_window; }
-int &Camera_View::get_height_window() { return _height_window; }
+float &Camera_View::get_width_window() { return _width_window; }
+float &Camera_View::get_height_window() { return _height_window; }
 scm::math::mat4f &Camera_View::get_matrix_perspective() { return _matrix_perspective; }
 void Camera_View::set_position(scm::math::vec3f position)
 {
