+++ apps/provis/Camera_View.h
@@ -58,8 +58,8 @@ class Camera_View
     scm::math::quat<float> _rotation = _rotation_initial;
     scm::math::mat4f _matrix_view;
     scm::math::mat4f _matrix_perspective;
-    int _width_window;
-    int _height_window;
+    float _width_window;
+    float _height_window;
 
     void update_matrices();
 
@@ -70,15 +70,15 @@ class Camera_View
     Camera_View();
 
     void translate(scm::math::vec3f offset);
-    void update_window_size(int width_window, int height_window);
+    void update_window_size(float width_window, float height_window);
     scm::math::mat4f &get_matrix_view();
     scm::math::mat4f &get_matrix_perspective();
 
     void set_position(scm::math::vec3f position);
     scm::math::vec3f &get_position();
     scm::math::quat<float> &get_rotation();
-    int &get_width_window();
-    int &get_height_window();
+    float &get_width_window();
+    float &get_height_window();
     void set_rotation(scm::math::quat<float> rotation);
     bool _mode_is_ego = true;
